fix: Refactoring and minor fixes for PR #49
